+++ classes/Flex.php
@@ -74,13 +74,7 @@ class Flex extends \Grav\Framework\Flex\Flex
 
         // Order blueprints by title.
         usort($directories, static function (FlexDirectory $a, FlexDirectory $b) {
-            $at = $a->getTitle();
-            $bt = $b->getTitle();
-            if ($at === $bt) {
-                return 0;
-            }
-
-            return $at < $bt ? -1 : 1;
+            return $a->getTitle() <=> $b->getTitle();
         });
 
         return $directories;